requirements.txt
bibtexparser
mistune
markdown-it-py
mdit-py-plugins
mdformat
watchdog
